+++ lib/super/form/schema_types.rb
@@ -74,6 +74,10 @@ module Super
         Generic.new(partial_path: "form_field_text", extras: extras, nested: {})
       end
 
+      def rich_text_area(**extras)
+        Generic.new(partial_path: "form_field_rich_text_area", extras: extras, nested: {})
+      end
+
       def has_many(reader, **extras)
         nested = @fields.nested do
           yield